Factor out benchmark parsing code to enable diffs more easily
